making configureStore compatible with Webpack 2
